refactor: extract row to separate component
