+++ src/components/table/row-renderer/row-renderer.js
@@ -2,6 +2,7 @@ import React, { PureComponent } from 'react'
 import { connect } from 'react-redux'
 import { singleRowSelected, multipleRowsSelected } from '../../../actions/actions'
 import TableHeaders from '../table-headers/table-headers'
+import Row from '../row/row'
 import store from '../../../store/store'
 
 class RowRenderer extends PureComponent {
@@ -16,43 +17,27 @@ class RowRenderer extends PureComponent {
 
 	render () {
 		if (this.props.index === 0) {
+			console.log(this.props.style)
 			return (
-				<div className="table-row table-header" style={this.props.style}>
-					<TableHeaders />
-				</div>
+				<TableHeaders style={this.props.style} />
 			)
 		}
 
-		const { selectedRows } = this.props
 		const el = this.props.data[this.props.index - 1]
-		const married = el.isMarried ? "Yes" : "No"
-		const classes = ["table-row"]
-
-		el.gender === 'Male' ? classes.push('male') : classes.push('female')
-		if (selectedRows.includes(el.id)) {
-			classes.push('active')
-		}
 
 		return (
-			<div style={this.props.style} 
-				className={classes.join(' ')} 
-				onClick={(e)=> this.handleSelectRow(e, el.id)}
-			>
-				<div className="number">{this.props.index}</div>
-				<div>{el.firstName}</div>
-				<div>{el.lastName}</div>
-				<div>{el.gender}</div>
-				<div>{married}</div>
-				<div>{el.points}</div>
-				<div>{el.shirtSize}</div>
-			</div>
+			<Row 
+				style={this.props.style} 
+				cls={["table-row"]}
+				onClick={(e) => this.handleSelectRow(e, el.id)} 
+				el={el}
+				index={this.props.index} />
 		)
 	}
 }
 
-const mapStateToProps = ({ selectedRows }) => {
+const mapStateToProps = () => {
 	return {
-		selectedRows,
 	}
 }
